requirements.txt
tqdm==4.11.2
numpy==1.12.1
Pillow==2.6.0
dlib==19.4.0
ImageHash==3.4
rectpack==0.1
ImageHash==3.4
pytesseract==0.1.6
selenium==3.4.1
networkx==1.11
